+++ packages/core/src/utils/misc.ts
@@ -18,8 +18,8 @@ export const numericRegex: RegExp = new RegExp(
  * Determines if a variable is a plain old JavaScript object, aka POJO.
  */
 // oxlint-disable-next-line typescript/no-explicit-any
-export const isPojo = (obj: any): obj is Record<string, any> =>
-  obj === null || typeof obj !== 'object' ? false : Object.getPrototypeOf(obj) === Object.prototype;
+export const isPojo = (obj: unknown): obj is Record<string, any> =>
+  obj !== null && typeof obj === 'object' && Object.getPrototypeOf(obj) === Object.prototype;
 
 /**
  * Simple helper to determine whether a value is null, undefined, or an empty string.